Validate module notations in capabilities
This commit prepares the verification that two modules with the same capability
are not found in the dependency graph, by first converting the module notations
into a proper `ModuleIdentifier` instance.

It adds tests to make sure that we can provide a valid notation for the modules,
and adds an internal interface to capabilities that will be used for resolution.